move StringFloatVecClose to testutils
